Removed CatalogService EFCore and Added MongoDb project
