--- services/catalog/src/EShopOnAbp.CatalogService.EntityFrameworkCore/Properties/AssemblyInfo.cs
@@ -1,2 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-[assembly:InternalsVisibleToAttribute("EShopOnAbp.CatalogService.EntityFrameworkCore.Tests")]
