+++ services/catalog/src/EShopOnAbp.CatalogService.MongoDB/MongoDB/ICatalogServiceMongoDbContext.cs
@@ -1,0 +1,13 @@
+﻿using Volo.Abp.Data;
+using Volo.Abp.MongoDB;
+
+namespace EShopOnAbp.CatalogService.MongoDB
+{
+    [ConnectionStringName(CatalogServiceDbProperties.ConnectionStringName)]
+    public interface ICatalogServiceMongoDbContext : IAbpMongoDbContext
+    {
+        /* Define mongo collections here. Example:
+         * IMongoCollection<Question> Questions { get; }
+         */
+    }
+}
