--- services/catalog/test/EShopOnAbp.CatalogService.EntityFrameworkCore.Tests/EntityFrameworkCore/CatalogServiceEntityFrameworkCoreTestBase.cs
@@ -1,9 +1,0 @@
-﻿using Volo.Abp;
-
-namespace EShopOnAbp.CatalogService.EntityFrameworkCore
-{
-    public abstract class CatalogServiceEntityFrameworkCoreTestBase : CatalogServiceTestBase<CatalogServiceEntityFrameworkCoreTestModule> 
-    {
-
-    }
-}
