--- services/catalog/test/EShopOnAbp.CatalogService.EntityFrameworkCore.Tests/EntityFrameworkCore/CatalogServiceEntityFrameworkCoreTestModule.cs
@@ -1,62 +1,0 @@
-﻿using Microsoft.Data.Sqlite;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.EntityFrameworkCore.Storage;
-using Microsoft.Extensions.DependencyInjection;
-using Volo.Abp;
-using Volo.Abp.EntityFrameworkCore;
-using Volo.Abp.EntityFrameworkCore.Sqlite;
-using Volo.Abp.Modularity;
-
-namespace EShopOnAbp.CatalogService.EntityFrameworkCore
-{
-    [DependsOn(
-        typeof(CatalogServiceEntityFrameworkCoreModule),
-        typeof(CatalogServiceTestBaseModule),
-        typeof(AbpEntityFrameworkCoreSqliteModule)
-        )]
-    public class CatalogServiceEntityFrameworkCoreTestModule : AbpModule
-    {
-        private SqliteConnection _sqliteConnection;
-
-        public override void ConfigureServices(ServiceConfigurationContext context)
-        {
-            ConfigureInMemorySqlite(context.Services);
-        }
-
-        private void ConfigureInMemorySqlite(IServiceCollection services)
-        {
-            _sqliteConnection = CreateDatabaseAndGetConnection();
-
-            services.Configure<AbpDbContextOptions>(options =>
-            {
-                options.Configure(context =>
-                {
-                    context.DbContextOptions.UseSqlite(_sqliteConnection);
-                });
-            });
-        }
-
-        public override void OnApplicationShutdown(ApplicationShutdownContext context)
-        {
-            _sqliteConnection.Dispose();
-        }
-
-        private static SqliteConnection CreateDatabaseAndGetConnection()
-        {
-            var connection = new SqliteConnection("Data Source=:memory:");
-            connection.Open();
-
-            var options = new DbContextOptionsBuilder<CatalogServiceDbContext>()
-                .UseSqlite(connection)
-                .Options;
-
-            using (var context = new CatalogServiceDbContext(options))
-            {
-                context.GetService<IRelationalDatabaseCreator>().CreateTables();
-            }
-
-            return connection;
-        }
-    }
-}
